TY - Commit No 7 - Game Master, Game State and Cursor
Created the Game Master object which will hold the GM, game master script, along with things like the audio manager. Scripts that only require one instance.
Created an enum to track the game sate used to control the cursor and in the future other elements between the Tour, UI and 2D Game.
Make the cursor become invisiable and locked to screen when moving around and come back with the click of escape. Also works with controllers.