TY - Commit No 8 - Footsteps and Mixers
Creates the multiple mixers for the project to allow the user to change the volume of music, sound and voice independantly.
Created the footstep function, required changing the movement script to only run when there was a magnitude to the input vector to stop footsteps playing when character wasn't moving. The footsteps are played by a counter that counts 1 per second but plays a footstep sound every 0.6seconds, can be adjusted along with the speed to better match a potential students stepwidth.